requirements.txt
wheel==0.37.1
configparser==5.0.2
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Admin==1.6.0
Flask-Security==3.0.0
psycopg2-binary==2.9.1
Flask-BabelEx==0.9.4
SQLAlchemy==1.4.25
email-validator==1.2.1
uWSGI==2.0.20
pyyaml==6.0
bcrypt==3.2.0
requests==2.27.1
Flask-Mail==0.9.1
